backend/ai_brain/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
gtts>=2.5.0
pyttsx3>=2.90